Refactor all views
Break up huge view files into smaller files
