+++ Applite/Utilities/Alert Manager/AlertManager.swift
@@ -56,32 +56,3 @@ final class AlertManager: ObservableObject {
         primaryAction = nil
     }
 }
-
-struct AlertModifier: ViewModifier {
-    @ObservedObject var manager: AlertManager
-
-    func body(content: Content) -> some View {
-        content
-            .alert(manager.title, isPresented: $manager.isPresented) {
-                Button(manager.primaryButtonTitle) {
-                    manager.primaryAction?()
-                    manager.dismiss()
-                }
-
-                // Add cancel button if we have a primary action
-                if manager.primaryAction != nil {
-                    Button("Cancel", role: .cancel) {
-                        manager.dismiss()
-                    }
-                }
-            } message: {
-                Text(manager.message)
-            }
-    }
-}
-
-extension View {
-    func alertManager(_ manager: AlertManager) -> some View {
-        modifier(AlertModifier(manager: manager))
-    }
-}
