+++ Applite/Views/Components/Brew Path Selector/BrewPathSelectorView+CustomPathOption.swift
@@ -1,0 +1,39 @@
+//
+//  BrewPathSelectorView+CustomPathOption.swift
+//  Applite
+//
+//  Created by Milán Várady on 2024.12.26.
+//
+
+import SwiftUI
+
+extension BrewPathSelectorView {
+    func customPathOption(option: BrewPaths.PathOption) -> some View {
+        VStack(alignment: .leading, spacing: 5) {
+            pathOption(option, showPath: false)
+
+            HStack {
+                TextField("Custom brew path", text: $customBrewPathDebounced.text, prompt: Text("/path/to/brew"))
+                    .textFieldStyle(.roundedBorder)
+                    .frame(maxWidth: 300)
+                    .autocorrectionDisabled()
+
+                Button("Browse") {
+                    choosingCustomFolder = true
+                }
+                .fileImporter(
+                    isPresented: $choosingCustomFolder,
+                    allowedContentTypes: [.unixExecutable]
+                ) { result in
+                    switch result {
+                    case .success(let file):
+                        customBrewPathDebounced.text = file.path
+                    case .failure(let error):
+                        print(error.localizedDescription)
+                    }
+                }
+            }
+            .disabled(brewPathOption != BrewPaths.PathOption.custom.rawValue)
+        }
+    }
+}
