+++ Applite/Views/Components/SmallProgressView.swift
