+++ Applite/Views/Components/Sparkle Updater/CheckForUpdatesView.swift
