+++ Applite/Views/Detail Views/Discover/Discover Section/DiscoverSectionView.swift
@@ -1,0 +1,26 @@
+//
+//  DiscoverSectionView.swift
+//  Applite
+//
+//  Created by Milán Várady on 2024.12.26.
+//
+
+import SwiftUI
+
+struct DiscoverSectionView: View {
+    let category: Category
+    @Binding var navigationSelection: String
+
+    @EnvironmentObject var caskData: CaskData
+
+    @State var scrollOffset: CGFloat = 0
+
+    var body: some View {
+        VStack(alignment: .leading) {
+            categoryHeader
+
+            appRowAndControls
+                .frame(height: AppView.dimensions.height * 2 + 20)
+        }
+    }
+}
