+++ Applite/Views/Detail Views/Download/DownloadView+SortingOptions.swift
@@ -1,0 +1,30 @@
+//
+//  DownloadView+SortingOptions.swift
+//  Applite
+//
+//  Created by Milán Várady on 2024.12.26.
+//
+
+import SwiftUI
+
+extension DownloadView {
+    var sortingOptions: some ToolbarContent {
+        ToolbarItem {
+            Menu {
+                Picker("Sort by", selection: $sortBy) {
+                    ForEach(SortingOptions.allCases) { option in
+                        Text(LocalizedStringKey(option.rawValue)).tag(option)
+                    }
+                }
+                .pickerStyle(.inline)
+                
+                Toggle(isOn: $hideUnpopularApps) {
+                    Text("Hide apps with few downloads")
+                }
+            } label: {
+                Label("Sorting Options", systemImage: "slider.horizontal.3")
+                    .labelStyle(.titleAndIcon)
+            }
+        }
+    }
+}
